# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'Clubhouse Web API'
  description: 'Web Resources Specification (A7) for Clubhouse'
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /login:
    get:
      operationId: R101
      summary: 'R101: Login Form'
      description: 'Provide login form. Access: PUB'
      tags:
        - 'M01: Authentication and Profiles'
      responses:
        '200':
          description: 'Ok. Show login UI.'
        '302':
          description: 'User is already logged in.'
          headers:
            Location:
              description: 'URL to redirect if access fails.'
              schema:
                type: string
                example: '/catalog'

    post:
      operationId: R102
      summary: 'R102: Login Action'
      description: 'Processes the login form submission. Access: PUB'
      tags:
        - 'M01: Authentication and Profiles'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                remember-me:
                    type: boolean
              required:
                - email
                - password
      responses:
        '302':
          description: 'Redirect after processing login credentials.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Authentication succeeded. Redirect to catalog page.'
                  value: '/catalog'
                302Error:
                  description: 'Authentication failed. Redirect to login form.'
                  value: '/login'

  # /logout
...